requirements.txt
moviepy
faster-whisper
litellm
python-dotenv
requests
pexels-api
pixabay-python
Pillow